move languageSupport dir
