+++ packages/components/.prettierignore
@@ -1,3 +1,3 @@
 dist/
 storybook-static
-src/languageSupport/generated/
+src/components/CodeEditor/languageSupport/generated/
