+++ packages/components/package.json
@@ -92,7 +92,7 @@
     "dev": "storybook dev -p 6006 --no-version-updates",
     "build:ts": "rm -rf dist/src dist/tsconfig.build.tsbuildinfo && tsc -b",
     "build:css": "postcss ./src/styles/full.css -o ./dist/full.css && postcss ./src/styles/common.css -o ./dist/common.css",
-    "build:lezer": "cd ./src/languageSupport; mkdir -p generated; lezer-generator ./squiggle.grammar --output generated/squiggle.ts",
+    "build:lezer": "cd ./src/components/CodeEditor/languageSupport; mkdir -p generated; lezer-generator ./squiggle.grammar --output generated/squiggle.ts",
     "build:storybook": "storybook build",
     "build": "pnpm run build:lezer && pnpm run build:ts && pnpm run build:css && pnpm run build:storybook",
     "lint": "pnpm lint:prettier && pnpm eslint",
