+++ packages/components/src/components/CodeEditor/languageSupport/highlightingStyle.ts
