+++ packages/components/src/components/CodeEditor/languageSupport/squiggle.grammar
