+++ packages/components/src/components/CodeEditor/languageSupport/squiggle.ts
