+++ packages/components/src/components/DynamicSquiggleViewer.tsx
@@ -3,7 +3,7 @@ import { forwardRef } from "react";
 import { SquiggleViewer } from "../index.js";
 import { SquiggleOutput } from "../lib/hooks/useSquiggle.js";
 import { getResultVariables, getResultValue } from "../lib/utility.js";
-import { CodeEditorHandle } from "./CodeEditor.js";
+import { CodeEditorHandle } from "./CodeEditor/index.js";
 import { PartialPlaygroundSettings } from "./PlaygroundSettings.js";
 import { SquiggleViewerHandle } from "./SquiggleViewer/index.js";
 import { ErrorBoundary } from "./ErrorBoundary.js";
