+++ packages/ui/.prettierignore
@@ -1,3 +1,2 @@
 dist/
 storybook-static
-src/languageSupport/generated/
